feat: restructure getGMTOffsetText
